+++ src/mainwindow.h
@@ -8,6 +8,7 @@
 #include <QSystemTrayIcon>
 
 class SipClient;
+class CallManager;
 class CallNotification;
 class ScrollHelper;
 
@@ -17,6 +18,10 @@ class MainWindow : public QMainWindow
 public:
     explicit MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
+
+    static QString parseNumber(const QString &uri);
+    static QString parseDisplayName(const QString &uri);
+    static QString formatDuration(int seconds);
 
 protected:
     void closeEvent(QCloseEvent *event) override;
@@ -34,7 +39,9 @@ private slots:
     void onRegisterClicked();
     void onRegistrationStatus(bool ok, const QString &message);
     void onCallStateChanged(int callId, const QString &state);
+    void onCallMediaActive(int callId);
     void onIncomingCall(int callId, const QString &remoteUri);
+    void onCallDurationChanged(int seconds);
 
     void onAbout();
     void onSettings();
@@ -42,12 +49,8 @@ private slots:
     void onTrayRestore();
     void onTrayExit();
     void updateClock();
-    void updateCallDuration();
 
 private:
-    static QString parseNumber(const QString &uri);
-    static QString parseDisplayName(const QString &uri);
-
     void setupMenu();
     void setupTray();
     void loadSettings();
@@ -57,6 +60,7 @@ private:
     void setCallDisplay();
 
     SipClient *m_sipClient;
+    CallManager *m_callManager;
 
     QSystemTrayIcon *m_trayIcon = nullptr;
     bool m_forceQuit = false;
@@ -74,11 +78,7 @@ private:
     QTimer *m_longPressTimer;
     QTimer *m_zeroTimer;
     QTimer *m_clockTimer;
-    QTimer *m_callDurationTimer;
-    int m_callDuration = 0;
 
-    bool m_inCall = false;
-    bool m_incomingWaiting = false;
     bool m_longPressFired = false;
     bool m_dialingMode = false;
     QString m_dialedNumber;
@@ -87,6 +87,8 @@ private:
     QString m_username;
     QString m_password;
     int m_port = 0;
+    bool m_echoCancel = true;
+    int m_echoAggressiveness = 1;
 
     CallNotification *m_callNotification = nullptr;
 };